Refactored the frontend part of the project Delete static folder Fixed build script for index.ejs
