+++ package.json
@@ -33,7 +33,7 @@
     "build:publish": "yarn build:clean && yarn build:ui && yarn build:routes --project tsconfig.publish.json",
     "build:ui": "NODE_ENV=production webpack",
     "build:ui:watch": "webpack --watch",
-    "build:routes": "yarn tsc",
+    "build:routes": "yarn tsc && cp ./src/ui/index.ejs dist/ui/index.ejs",
     "build:clean": "rm -rf ./static ./dist",
     "start:dev:docker": "docker-compose up -d",
     "start:dev:server": "nodemon example.js",
