+++ src/ui/index.css
@@ -196,7 +196,6 @@ section {
 
 button {
   background: #333;
-  margin: 10px;
   padding: 12px 12px;
   cursor: pointer;
   font-weight: 600;
